# openapi.yaml
openapi: 3.0.3
info:
  title: Cyber Storefront API
  version: "1.0.0"
  description: |
    **Cyber** ონლაინ მაღაზიის API: კატალოგი, ავტორიზაცია, კალათა, ჩექაუთი და პროფილი.

    ---

    # 🚀 სწრაფი დაწყება — სცადე აქვე, 4 ნაბიჯად

    ამ გვერდზე ყველა ენდპოინტის გატესტვა შეგიძლია, კოდის დაწერის გარეშე.

    ### ნაბიჯი 1 — შექმენი ანგარიში

    ჩამოდი `POST /auth/register`-მდე → დააჭირე **`Try it out`** (მარჯვნივ)
    → შეცვალე მონაცემები → **`Execute`**

    ```json
    { "name": "Test User", "email": "შენი@example.com", "password": "Password1" }
    ```

    ### ნაბიჯი 2 — დააკოპირე ტოკენი

    პასუხში (Response body) ნახავ:

    ```json
    {
      "user": { "id": "...", "name": "Test User", "email": "..." },
      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOi..."
    }
    ```

    დააკოპირე **`accessToken`-ის მნიშვნელობა** — ბრჭყალების გარეშე.

    ### ნაბიჯი 3 — დააჭირე `Authorize` 🔓

    ამ გვერდის **ზედა მარჯვენა კუთხეში** არის ღილაკი **`Authorize`** (ბოქლომის ხატულით).

    1. დააჭირე
    2. ველში `Value` ჩასვი დაკოპირებული ტოკენი
       *(მხოლოდ ტოკენი — სიტყვა `Bearer` ჩაწერა არ გჭირდება, Swagger თავად დაამატებს)*
    3. **`Authorize`** → **`Close`**

    ბოქლომები 🔒 დაიკეტება — ესე იგი ავტორიზებული ხარ.

    ### ნაბიჯი 4 — სცადე დაცული ენდპოინტი

    ჩამოდი `GET /auth/me`-მდე → **`Try it out`** → **`Execute`**

    ✅ დაბრუნდება შენი მომხმარებლის მონაცემები.

    **სცადე საწინააღმდეგოც:** დააჭირე `Authorize` → **`Logout`** → ისევ გაუშვი
...

    ---

...

...
servers:
...
paths:
...
  /checkout:
    get:
      tags: [Checkout]
      summary: ჩექაუთის გვერდის მონაცემები
      description: |
        🔒 **ჩექაუთის გვერდი ავტორიზაციის გარეშე არ იხსნება** — `401`-ზე გადაამისამართე `/login`-ზე
        და შესვლის შემდეგ დააბრუნე ჩექაუთზე.

        აბრუნებს კალათას, ჯამებს მიწოდების ჩათვლით და `shipping`-ს — პროფილიდან
        წინასწარ შევსებულ მისამართს ფორმისთვის. ცარიელ კალათაზე `items: []`.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /orders:
    post:
      tags: [Checkout]
      summary: "გადახდა — ნაბიჯი 1: შეკვეთა + ბარათი → კოდი ელფოსტაზე"
      description: |
        🔒 შეკვეთა იქმნება **სერვერის კალათიდან** (პროდუქტების სიას თავად არ აგზავნი) სტატუსით
        `pending_payment`, ხოლო ელფოსტაზე მიდის 6-ციფრიანი კოდი (ვადა **10 წუთი**).

        კოდი, როცა `DEV_EXPOSE_RESET_CODE=true`, ბრუნდება `devCode`-შიც.

        სატესტო ბარათები: `4242 4242 4242 4242` ✅, `4000 0000 0000 0002` → `402 CARD_DECLINED`.
        ვალიდაციის შეცდომები ჩადგმულ ველებზე: `errors["card.number"]`, `errors["shipping.phone"]`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shipping, card]
              properties:
                shipping: { $ref: "#/components/schemas/Shipping" }
                card:
                  type: object
                  required: [number, holder, expiry, cvc]
                  properties:
                    number: { type: string, example: "4242 4242 4242 4242" }
                    holder: { type: string, example: "NINO BERIDZE" }
                    expiry: { type: string, pattern: "^(0[1-9]|1[0-2])/\\d{2}$", example: "12/29" }
                    cvc: { type: string, example: "123" }
      responses:
        "201":
          description: შეკვეთა შეიქმნა, კოდი გაიგზავნა
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId: { type: string }
                  status: { type: string, example: pending_payment }
                  total: { type: number, example: 245.5 }
                  currency: { type: string, example: GEL }
                  message: { type: string }
                  expiresInMinutes: { type: integer, example: 10 }
                  devCode: { type: string, example: "483920", description: "მხოლოდ DEV_EXPOSE_RESET_CODE=true-ზე" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402":
          description: ბარათი უარყოფილია (`CARD_DECLINED`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: კალათა ცარიელია (`CART_EMPTY`) ან მარაგი არ კმარა (`INSUFFICIENT_STOCK`, `items[]`-ით)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/Validation"
    get:
      tags: [Checkout]
      summary: ჩემი შეკვეთები
      description: 🔒 შეკვეთების ისტორია, ახალი პირველი.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Order" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /orders/{id}/confirm:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [Checkout]
      summary: "გადახდა — ნაბიჯი 2: კოდის დადასტურება → გადახდა"
      description: |
        🔒 სწორ კოდზე შეკვეთა ხდება `paid`, მარაგი ჩამოიჭრება და **კალათა სუფთავდება**.
        5 არასწორი მცდელობის შემდეგ — `429`, მოითხოვე ახალი კოდი.

        თუ კოდის მოსვლამდე ვიღაცამ ბოლო ნივთი იყიდა — `409 INSUFFICIENT_STOCK`, არაფერი ჩამოიჭრება.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, pattern: "^\\d{6}$", example: "483920" }
      responses:
        "200":
          description: გადახდილია
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: "#/components/schemas/Order" }
        "400":
          description: კოდი არასწორი ან ვადაგასულია (`INVALID_PAYMENT_CODE`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OrderNotFound" }
        "409":
          description: უკვე გადახდილია (`ORDER_ALREADY_PAID`) ან მარაგი აღარ კმარა (`INSUFFICIENT_STOCK`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/Validation"
        "429":
          description: ძალიან ბევრი მცდელობა (`TOO_MANY_ATTEMPTS`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /orders/{id}/resend-code:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      tags: [Checkout]
      summary: კოდის ხელახლა გაგზავნა
      description: 🔒 ახალი კოდი, ახალი 10 წუთი, მცდელობები ნულდება. **30 წამში ერთხელ.**
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: გაიგზავნა
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  expiresInMinutes: { type: integer, example: 10 }
                  devCode: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OrderNotFound" }
        "409":
          description: შეკვეთა უკვე გადახდილია (`ORDER_ALREADY_PAID`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: ძალიან ადრეა (`RESEND_TOO_SOON`, `retryAfterSeconds`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /orders/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Checkout]
      summary: ერთი შეკვეთა
      description: 🔒 მხოლოდ საკუთარი — სხვისი შეკვეთის id-ზე `404`. შეკვეთის დადასტურების გვერდისთვის.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OrderNotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: ტოკენი არ არის, არასწორია ან ვადაგასულია (`UNAUTHORIZED` / `TOKEN_EXPIRED`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    OrderNotFound:
      description: შეკვეთა არ მოიძებნა (`ORDER_NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Validation:
      description: ვალიდაციის შეცდომა (`VALIDATION_ERROR`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
...
    CartItem:
      type: object
      properties:
        id: { type: string, description: "პოზიციის id — PATCH/DELETE-ისთვის" }
        productId: { type: string }
        qty: { type: integer, example: 2 }
        lineTotal: { type: number, example: 199.98 }
        product:
          type: object
          properties:
            id: { type: string }
            slug: { type: string }
            title: { type: string }
            brand: { type: string }
            image: { type: string }
            price: { type: number, example: 99.99 }
            oldPrice: { type: number, nullable: true }
            currency: { type: string, example: GEL }
            stock: { type: integer, example: 12 }
            inStock: { type: boolean }
...
    Shipping:
      type: object
      required: [fullName, phone, city, address]
      properties:
        fullName: { type: string, example: "Nino Beridze" }
        phone: { type: string, example: "+995 555 12 34 56" }
        city: { type: string, example: "Tbilisi" }
        address: { type: string, example: "Rustaveli Ave 12, apt 5" }
    CheckoutSummary:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CartItem" } }
        totalQty: { type: integer }
        subtotal: { type: number, example: 89.5 }
        shippingFee: { type: number, example: 5 }
        total: { type: number, example: 94.5 }
        currency: { type: string, example: GEL }
        freeShippingFrom: { type: number, example: 100 }
        shipping: { $ref: "#/components/schemas/Shipping" }
    Order:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [pending_payment, paid] }
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string, nullable: true, description: "null, თუ პროდუქტი კატალოგიდან წაიშალა" }
              slug: { type: string }
              title: { type: string }
              image: { type: string }
              price: { type: number, description: "ფასი შეკვეთის მომენტში" }
              qty: { type: integer }
              lineTotal: { type: number }
        subtotal: { type: number }
        shippingFee: { type: number }
        total: { type: number }
        currency: { type: string, example: GEL }
        shipping: { $ref: "#/components/schemas/Shipping" }
        payment:
          type: object
          properties:
            brand: { type: string, example: visa }
            last4: { type: string, example: "4242" }
        createdAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time, nullable: true }
...
    Error:
      type: object
      properties:
        message: { type: string, example: "Invalid email or password" }
        code: { type: string, example: "INVALID_CREDENTIALS" }
    ValidationError:
      type: object
      properties:
        message: { type: string, example: "Validation failed" }
        code: { type: string, example: "VALIDATION_ERROR" }
        errors:
          type: object
          additionalProperties: { type: string }
          example: { email: "Enter a valid email address", password: "Password must be at least 8 characters" }